Fix retry options share
