+++ lib/file.js
@@ -23,6 +23,7 @@ const { BlockDownload } = require('./block/blockdownload');
 const { BlockUpload } = require('./block/blockupload');
 
 const RETRY_OPTIONS_FILTER = [
+    'timeout',
     'retryMaxDuration', 'retryBackoff', 'retryInitialDelay',
     'retryEnabled', 'retryAllErrors', 'retryMaxCount'
 ];
@@ -142,7 +143,7 @@ async function downloadFileConcurrently(url, filepath, options) {
     // extract upload options
     const downloadOptions = filterObject(
         options || {},
-        ['timeout', 'headers', ...CONCURRENCY_OPTIONS_FILTER, ...RETRY_OPTIONS_FILTER, 'preferredPartSize']
+        ['headers', ...CONCURRENCY_OPTIONS_FILTER, ...RETRY_OPTIONS_FILTER, 'preferredPartSize']
     );
 
     return downloader.downloadFiles({
@@ -186,7 +187,7 @@ async function uploadFileConcurrently(filepath, url, options) {
     // extract upload options
     const uploadOptions = filterObject(
         options || {},
-        ['method', 'timeout', 'headers', ...CONCURRENCY_OPTIONS_FILTER,
+        ['method', 'headers', ...CONCURRENCY_OPTIONS_FILTER,
             ...RETRY_OPTIONS_FILTER]
     );
 
@@ -250,7 +251,7 @@ async function uploadMultiPartFileConcurrently(filepath, target, options = {}) {
     // extract upload options
     const uploadOptions = filterObject(
         options || {},
-        ['method', 'timeout', 'headers', ...CONCURRENCY_OPTIONS_FILTER,
+        ['method', 'headers', ...CONCURRENCY_OPTIONS_FILTER,
             ...RETRY_OPTIONS_FILTER, 'preferredPartSize']
     );
